add function for workTime

+++ backend/src/main/java/com/example/backend/model/monthly/DailyPlan.java
@@ -15,5 +15,5 @@ public class DailyPlan {
     LocalTime start;
     LocalTime end;
     List<Pause> pause;
-    Double time;
+    long time;
 }
